refactor: split telegram_callbacks.py into focused modules (#203)
